Removed deprecated functions
